+++ src/Zope2/Startup/handlers.py
@@ -57,12 +57,6 @@ def large_file_threshold(value):
     import ZServer
     ZServer.LARGE_FILE_THRESHOLD = value
 
-def publisher_profile_file(value):
-    value is not None and _setenv('PROFILE_PUBLISHER', value)
-    from ZPublisher.Publish import install_profiling
-    install_profiling(value)
-    return value
-
 def http_realm(value):
     value is not None and _setenv('Z_REALM', value)
     return value